feat: Add authentication check for game cart addition
This commit introduces an authentication check before a user can add a game to the cart. If the user is not authenticated, they are prompted to log in before they can place an order. Changes include addition of Redux state selector for authentication status, and conditional rendering of cart addition button based on authentication status. Also, some CSS styles have been added for new elements.