feat: add navigation link to shop button on home page
In this commit, the "Jetzt einkaufen" button on the home page has been linked to the games page using NavLink from react-router-dom. This provides the user with a direct navigation to the games page. Additionally, a margin has been added to the hero section in the CSS for better layout.